user can now export files in the directory as a txt file

+++ tk1.py
@@ -6,28 +6,70 @@ root = tk.Tk()
 root.geometry('200x400')
 root.minsize(400,400)
 root.maxsize(400,800)
-root.title('Y2Mate rename')
+root.title('File rename')
+
+global substring
+global file_list
+global rename_count
+global dir
+
+substring = 'y2mate.com - '
+file_list = []
+rename_count = 0
+dir = './'
 
 # lists all files in the selected directory
 def check_dir():
+    eligible_files = 0
     global file_list
-    global dir
     file_list = os.listdir(dir)
-    print(f'list of files in directory: {file_list}\n')
-    lbl2.config(text = str(file_list))
 
+    for file in file_list:
+        if substring in file:
+            eligible_files += 1
+
+    export_list_as_txt()
+
+    print(f'list of files in directory: {file_list}\n')
+    lbl2.config(text = f'{eligible_files} file(s) eligible for renaming')
+
+def export_list_as_txt():
+    txt_name = './files_list.txt'
+    count = 0
+
+    while True:
+        if os.path.exists(txt_name):
+            print(f'{txt_name} already exists, changing name')
+            txt_name = f'./files_list_{count}.txt'
+            count += 1
+        else:
+            break
+
+    f = open(txt_name, 'a', encoding='utf-8')
+
+    if not file_list:
+        print('No files')
+        return
+    
+    for item in file_list:
+        f.write(item + '\n')
+
+# renames files in directory
 def rename_files():
     label_msg = ""
-    global file_list
-    global substring
-    global rename_count
-    global dir
+    rename_count = 0
 
     for file_name in file_list:
         if substring in file_name:
             new_name = os.path.join(dir, file_name[13:])
             old_name = os.path.join(dir, file_name)
-            os.rename(old_name, new_name)
+
+            if not os.path.exists(new_name):
+                os.rename(old_name, new_name)
+            else:
+                print(f"File '{new_name}' already exists. Skipping renaming.")
+                continue
+
             print(f'Renamed {old_name} to {new_name}')
             label_msg += f'Renamed {old_name} to {new_name} \n'
             rename_count += 1
@@ -36,6 +78,7 @@ def rename_files():
     messagebox.showinfo("Done", f"Renamed {rename_count} files")
     lbl2.config(text = label_msg)
 
+# changes directory where renaming happens
 def change_directory():
     global dir
     new_dir = filedialog.askdirectory(initialdir= '/',
@@ -46,14 +89,9 @@ def change_directory():
         lbl1.config(text=f'Current directory: {dir}')
 
 
-## Variables ##
-file_list = []
-substring = 'y2mate.com - '
-rename_count = 0
-dir = './'
-
-
-## Widgets ##
+###
+### Widgets 
+###
 lbl1 = tk.Label(root, 
                 text=f'Current directory: {dir}')
 lbl1.pack()
@@ -64,7 +102,7 @@ dirSelectBtn = tk.Button(root,
 dirSelectBtn.pack()
 
 checkBtn = tk.Button(root, 
-                     text = 'Check folder', 
+                     text = 'Check directory', 
                      command = check_dir)
 checkBtn.pack()
